# [RFC] action/playbook: Add turbo mode for faster local execution