--- a/lib/ansible/modules/setup.py
+++ b/lib/ansible/modules/setup.py
@@ -145,7 +145,7 @@ EXAMPLES = """
 """
 
 # import module snippets
-from ..module_utils.basic import AnsibleModule
+from ansible.module_utils.basic import AnsibleModule
 
 from ansible.module_utils.facts.namespace import PrefixFactNamespace
 from ansible.module_utils.facts import ansible_collector