--- a/test/sanity/ignore.txt
+++ b/test/sanity/ignore.txt
@@ -135,6 +135,7 @@ lib/ansible/playbook/collectionsearch.py required-and-default-attributes  # http
 lib/ansible/playbook/helpers.py pylint:blacklisted-name
 lib/ansible/playbook/role/__init__.py pylint:blacklisted-name
 lib/ansible/plugins/action/normal.py action-plugin-docs # default action plugin for modules without a dedicated action plugin
+lib/ansible/plugins/action/turbo.py action-plugin-docs # experiment, likely wrongly placed
 lib/ansible/plugins/cache/base.py ansible-doc!skip  # not a plugin, but a stub for backwards compatibility
 lib/ansible/plugins/lookup/sequence.py pylint:blacklisted-name
 lib/ansible/plugins/strategy/__init__.py pylint:blacklisted-name