##### SUMMARY

This is a proof-of-concept / request-for-comments for speeding up local
Ansible execution. Caveat: It's very limited in scope and not generically
applicable as is (e.g., it ignores any details about become/runas usage).
The motivation for this experiment comes from an environment where fairly
basic playbooks using Ansible builtin functionality are run as root in
local mode.

It introduces a "turbo" module  style which, when the connection is `local`,
executes modules directly inside a subprocess instead of going through
AnsiballZ. In experiments, this improves playbook runtime by a factor
of 3 to 5.



##### ISSUE TYPE
- Feature Pull Request

##### COMPONENT NAME
action/playbook

##### ADDITIONAL INFORMATION

A simple performance experiment was done with the following playbook:

    ---
    - name: Test
      hosts: localhost
      gather_facts: False
      tasks:
      - template:
          src: 'template1.j2'
          dest: '/tmp/template-item-{{ item }}'
        loop: '{{ range(32) }}'
      - copy:
          content: 'copy {{ item|string }}'
          dest: '/tmp/file_{{ item }}'
        loop: '{{ range(32) }}'

    $ time PYTHONPATH=$(pwd)/lib python3 bin/ansible-playbook ./perftest/test.yml
    ...
    real    0m24,434s
    user    0m21,394s
    sys     0m4,843s

    $ time PYTHONPATH=$(pwd)/lib python3 bin/ansible-playbook ./perftest/test.yml --turbo
    ...
    real    0m4,707s
    user    0m3,519s
    sys     0m1,909s

This was added without any real knowledge of the Ansible code base, so it
may look somewhat off ;-)

The purpose of this is mainly to to hear if maybe there are other ways
to speed up local Ansible execution and/or whether integrating a real
version of something like this could be of interest.