@Akasurde 2020-10-16
@awelzel Thanks for submitting this feature. I would suggest to add this to [`community.general`](https://github.com/ansible-collections/community.general) collection since we are not adding new things directly in `ansible/ansible`. 

Also, there is another Turbo mode already available in [`cloud.common`](https://github.com/ansible-collections/cloud.common) which is being used by VMware REST modules. It would be great if you can take a look at that version as well.

Thanks.