@sivel 2020-10-20
Thank you for your interest in Ansible. I like where this is coming from, but we're going to pass on accepting this, or any specific variations on this implementation.

You may want to follow along with some development we are planning around short circuiting the forking mechanism for more simple tasks.

If you have further questions please stop by IRC or the mailing list:

   * IRC: #ansible on irc.freenode.net
   * mailing list: https://groups.google.com/forum/#!forum/ansible-project