@retr0h 2020-10-27
> You may want to follow along with some development we are planning around short circuiting the forking mechanism for more simple tasks.

Hey @sivel can you provide more information on this? Where can I find details on this?