@sivel 2020-10-27
We don't have a public document yet, although there have been a few proof of concept PRs that are working toward our goal.

Our near term plans are moving work out of the fork, and into the strategy, effectively short circuiting forking.

* https://github.com/ansible/ansible/pull/70477
* https://github.com/ansible/ansible/pull/71700
* https://github.com/ansible/ansible/pull/72057

We also hope to move looping to the strategy, rewriting the interface for process models, and switching to a different means of forking via concurrent.futures or asyncio. From there the plans keep expanding to enable us to make additional performance improvements, such as persistent workers, instead of a new worker per task.