@nitzmahone 2020-10-27
To add a little more color: our current worker architecture limits a number of "interesting" possibilities for major improvements to local and remote scale and performance; we've got some projects in flight that are removing the blockers to correcting that. WRT this problem in particular, there's actually a much easier and safer way to do what you're doing here that I've already been working with some folks to implement for the networking modules- it could pretty easily be extended to all local executions, but I wouldn't want to make it the default without a way for modules to opt-out first.