@awelzel 2020-10-28
Hey guys, thanks for all the input! It's good to see there's discussions around improving.

> WRT this problem in particular, there's actually a much easier and safer way to do what you're doing here that I've already been working with some folks to implement for the networking modules- it could pretty easily be extended to all local executions, [...]

@nitzmahone , do you have some pointers regarding this? Maybe there's a fork/branch/pr?

Thanks!